Issue #759 : Run permissions configurations - refactoring
